requirements.txt
Click==7.0
Flask==1.1.1
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.1
simpleaudio==1.0.2
Werkzeug==0.16.0
